Fix advanced oversight module initialization
