+++ index.html
@@ -90,6 +90,7 @@
     <script src="src/js/projects/TerraformingDurationProject.js"></script>
     <script src="src/js/projects/ScannerProject.js"></script>
     <script src="src/js/projects/SpaceMiningProject.js"></script>
+    <script src="src/js/projects/SpaceMirrorAdvancedOversight.js"></script>
     <script src="src/js/projects/SpaceMirrorFacilityProject.js"></script>
     <script src="src/js/projects/SpaceExportBaseProject.js"></script>
     <script src="src/js/projects/SpaceExportProject.js"></script>
